requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0  
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-rest-swagger==2.2.0 
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2       
pyaml==21.10.1
python-dotenv==1.0.0       
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
simplejson==3.18.4
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.3.0